+++ src/similar_reaction_combining_method.c
@@ -192,7 +192,7 @@ static int _CompareSpeciesEdge( CADDR_T e1, CADDR_T e2 ) {
     s1 = GetSpeciesInIREdge( (IR_EDGE*)e1 );
     s2 = GetSpeciesInIREdge( (IR_EDGE*)e2 );
             
-    return (int)s1 - (int)s2;
+    return (s1 - s2);
 }
 
 static KINETIC_LAW *_GenerateCombinedKineticLaw( KINETIC_LAW *kineticLaw1, KINETIC_LAW *kineticLaw2 ) {
